requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
coverage==6.4.4
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1.1
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-crontab==0.7.1
django-debug-toolbar==3.6.0
django-filter==22.1
django-haystack==3.2.1
django-import-export==2.9.0
django-jazzmin==2.5.0
django-matomo==0.1.6
django-recaptcha==3.0.0
django-redis==5.2.0
django-user-agents==0.4.0
djangorestframework==3.13.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fabric==2.7.1
fake-useragent==0.1.11
Faker==14.2.0
fastjsonschema==2.16.1
fontawesomefree==6.2.0
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
invoke==1.7.1
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
lxml==4.9.1
Markdown==3.4.1
markdown2==2.4.3
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.29
mysql-to-sqlite3==1.4.14
mysqlclient==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip-review==1.2.0
pipdeptree==2.3.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
Pygments==2.12.0
pymdown-extensions==9.5
PyNaCl==1.5.0
pyparsing==3.0.9
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qcloudsms-py==0.1.4
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.1.0
rcssmin==1.1.0
readtime==2.0.0
redis==4.3.4
requests==2.28.1
rjsmin==1.2.0
scipy==1.9.1
Send2Trash==1.8.0
sh==1.14.3
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlite3-to-mysql==1.4.16
sqlparse==0.4.2
stack-data==0.3.0
tablib==3.2.1
tabulate==0.8.10
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
ua-parser==0.15.0
Unidecode==1.3.4
urllib3==1.26.10
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
widgetsnbextension==3.6.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
